Update README with game outcome information
This commit adds a new section to the README titled "Game Outcome" that explains how the winner of the Rock Paper Scissors game will be recorded in the console after the completion of five rounds. This provides clarity to users about how the game results will be displayed. Additionally, minor formatting improvements were made for readability.